requirements.txt
torch
torchvision
opencv-python
scikit-learn
streamlit
matplotlib
numpy
Pillow
tqdm
pandas
fastapi
uvicorn
